+++ examples/plot_toy_model.py
@@ -16,15 +16,17 @@ from mapie.metrics import coverage_score
 regressor = LinearRegression()
 X, y = make_regression(n_samples=500, n_features=1, noise=20, random_state=59)
 
-mapie = MapieRegressor(regressor, method="jackknife_plus")
+mapie = MapieRegressor(regressor)
 mapie.fit(X, y)
 y_preds = mapie.predict(X)
 
-plt.xlabel('x')
-plt.ylabel('y')
+plt.xlabel("x")
+plt.ylabel("y")
 plt.scatter(X, y, alpha=0.3)
-plt.plot(X, y_preds[:, 0], color='C1')
+plt.plot(X, y_preds[:, 0], color="C1")
 order = np.argsort(X[:, 0])
 plt.fill_between(X[order].ravel(), y_preds[:, 1][order], y_preds[:, 2][order], alpha=0.3)
-plt.title(f"Target coverage = 0.9; Effective coverage = {coverage_score(y, y_preds[:, 1], y_preds[:, 2])}")
+plt.title(
+    f"Target coverage = 0.9; Effective coverage = {coverage_score(y, y_preds[:, 1], y_preds[:, 2])}"
+)
 plt.show()